pyproject.toml
[tool.poetry]
name = "obsidian-md-bot"
version = "0.1.0"
description = "Un'app Gradio per il webscraping e la creazione automatica di note Obsidian"
authors = ["Luca Baroncini"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
gradio = "^5.22.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
markdown = "^3.5.0"
transformers = "^4.35.0"
torch = ">=2.0.0"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
anthropic = "^0.20.0"
google-generativeai = "^0.3.2"
sentence-transformers = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
obsidian-md-bot = "src.main:main"